+++ packages/typespec-ts/src/transform/transfromRLCOptions.ts
@@ -92,6 +92,7 @@ function extractRLCOptions(
     emitterOptions["ignore-enum-member-name-normalize"];
   const compatibilityQueryMultiFormat =
     emitterOptions["compatibility-query-multi-format"];
+  const enableStorageCompat = emitterOptions["enable-storage-compat"] === true;
   const typespecTitleMap = emitterOptions["typespec-title-map"];
   const hasSubscriptionId = getSubscriptionId(dpgContext);
   const ignoreNullableOnOptional = getIgnoreNullableOnOptional(
@@ -136,7 +137,8 @@ function extractRLCOptions(
     hasSubscriptionId,
     ignoreNullableOnOptional,
     wrapNonModelReturn,
-    isMultiService
+    isMultiService,
+    enableStorageCompat
   };
 }